+++ Changelog.md
@@ -1,4 +1,7 @@
 # Changelog
+## [0.3.5]
+### Fixed
+    * Block toolbar buttons if list of plan, product, run of result-set is updated
 ## [0.3.4]
 ### Fixed
     * Fix errors after trying to edit ame of product to exists name